+++ lib/admin.php
@@ -14,7 +14,6 @@ function mai_admin_enqueue_scripts() {
 	// Add an editor stylesheet
 	add_editor_style( '/assets/css/mai-editor{$suffix}.css' );
 
-	// Register for later
 	wp_enqueue_style( 'mai-admin', MAI_PRO_ENGINE_PLUGIN_URL . "/assets/css/mai-admin{$suffix}.css", array(), MAI_PRO_ENGINE_VERSION );
 	wp_enqueue_script( 'mai-admin', MAI_PRO_ENGINE_PLUGIN_URL . "/assets/js/mai-admin{$suffix}.js", array( 'jquery' ), MAI_PRO_ENGINE_VERSION, true );
 }
@@ -90,6 +89,11 @@ function mai_login_logo_css() {
 			text-indent: 0 !important;
 			padding: 0 !important;
 			margin: 0 !important;
+		}
+		.login h1 img {
+			display: block !important;
+			max-width: 100% !important;
+			height: auto !important;
 		}
 	</style>';